requirement.txt
click==6.7
cmake==0.9.0
colorama==0.3.3
company==0.0.4
diff-match-patch==20121119
Django==1.11.8
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-classy-tags==0.8.0
django-cms==3.4.2
django-crispy-forms==1.6.0
django-filter==1.1.0
django-formtools==2.0
django-guardian==1.4.9
django-import-export==0.5.1
django-multiselectfield==0.1.8
django-reversion==2.0.10
django-sekizai==0.10.0
django-simple-menu==1.2.1
django-treebeard==4.1.0
django-user-accounts==1.3.1
djangocms-admin-style==1.2.6.2
djangorestframework==3.5.3
dlib==19.10.0
dukpy==0.2.2
et-xmlfile==1.0.1
face-recognition==1.2.1
face-recognition-models==0.3.0
future==0.16.0
httplib2==0.10.3
javascripthon==0.10
jdcal==1.3
Jinja2==2.10
jsonfield==2.0.1
jupyter-echarts-pypkg==0.1.2
lml==0.0.2
macropy3==1.1.0b2
Markdown==2.6.9
MarkupSafe==1.1.0
mysqlclient==1.3.9
numpy==1.14.2
odfpy==1.3.5
olefile==0.44
opencv-python==3.4.0.12
openpyxl==2.4.9
pickle-warehouse==0.1.1
Pillow==4.0.0
Pinax==0.9a2
pinax-cli==1.0.0
pinax-eventlog==1.1.2
pinax-theme-bootstrap==7.10.1
pinax-webanalytics==2.0.4
pycrypto==2.6.1
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pymssql==2.1.4
PyMySQL==0.7.9
pyodbc==4.0.19
pypinyin==0.35.1
PyQt5==5.10.1
pyqt5-tools==5.9.0.1.2 
pytz==2016.10
pywin32==223
PyYAML==3.12
requests==2.7.0
scipy==1.0.1
simplejson==3.16.0
sip==4.19.8
six==1.10.0
tablib==0.12.1
tzlocal==1.3
unicodecsv==0.14.1
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
wechat-sdk==0.6.4
win32core==221.36
win32gui==221.6
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.9.2
